# moderation_sdk/moderation_sdk/models/stats_response.py
# coding: utf-8

"""
    AI-Assisted Content Moderation & Publishing API

    Backend service for drafting, reviewing, and publishing blog posts.

    The version of the OpenAPI document: 1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json

from pydantic import BaseModel, ConfigDict, StrictFloat, StrictInt
from typing import Any, ClassVar, Dict, List, Union
from moderation_sdk.models.reason_count import ReasonCount
from moderation_sdk.models.status_count import StatusCount
from moderation_sdk.models.timeline_entry import TimelineEntry
from typing import Optional, Set
from typing_extensions import Self
from pydantic_core import to_jsonable_python

class StatsResponse(BaseModel):
    """
    StatsResponse
    """ # noqa: E501
    status_distribution: List[StatusCount]
    timeline: List[TimelineEntry]
    top_flagged_reasons: List[ReasonCount]
    total_posts: StrictInt
    approval_rate: Union[StrictFloat, StrictInt]
    flag_rate: Union[StrictFloat, StrictInt]
    __properties: ClassVar[List[str]] = ["status_distribution", "timeline", "top_flagged_reasons", "total_posts", "approval_rate", "flag_rate"]

    model_config = ConfigDict(
        validate_by_name=True,
        validate_by_alias=True,
        validate_assignment=True,
        protected_namespaces=(),
    )


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        return json.dumps(to_jsonable_python(self.to_dict()))

    @classmethod
    def from_json(cls, json_str: str) -> Optional[Self]:
        """Create an instance of StatsResponse from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        excluded_fields: Set[str] = set([
        ])

        _dict = self.model_dump(
            by_alias=True,
            exclude=excluded_fields,
            exclude_none=True,
        )
        # override the default output from pydantic by calling `to_dict()` of each item in status_distribution (list)
        _items = []
        if self.status_distribution:
            for _item_status_distribution in self.status_distribution:
                if _item_status_distribution:
                    _items.append(_item_status_distribution.to_dict())
            _dict['status_distribution'] = _items
        # override the default output from pydantic by calling `to_dict()` of each item in timeline (list)
        _items = []
        if self.timeline:
            for _item_timeline in self.timeline:
                if _item_timeline:
                    _items.append(_item_timeline.to_dict())
            _dict['timeline'] = _items
        # override the default output from pydantic by calling `to_dict()` of each item in top_flagged_reasons (list)
        _items = []
        if self.top_flagged_reasons:
            for _item_top_flagged_reasons in self.top_flagged_reasons:
                if _item_top_flagged_reasons:
                    _items.append(_item_top_flagged_reasons.to_dict())
            _dict['top_flagged_reasons'] = _items
        return _dict

    @classmethod
    def from_dict(cls, obj: Optional[Dict[str, Any]]) -> Optional[Self]:
        """Create an instance of StatsResponse from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "status_distribution": [StatusCount.from_dict(_item) for _item in obj["status_distribution"]] if obj.get("status_distribution") is not None else None,
            "timeline": [TimelineEntry.from_dict(_item) for _item in obj["timeline"]] if obj.get("timeline") is not None else None,
            "top_flagged_reasons": [ReasonCount.from_dict(_item) for _item in obj["top_flagged_reasons"]] if obj.get("top_flagged_reasons") is not None else None,
            "total_posts": obj.get("total_posts"),
            "approval_rate": obj.get("approval_rate"),
            "flag_rate": obj.get("flag_rate")
        })
        return _obj