# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI-Assisted Content Moderation & Publishing API",
    "description": "Backend service for drafting, reviewing, and publishing blog posts.",
    "version": "1.0.0"
  },
  "paths": {
    "/posts/stats": {
      "get": {
        "tags": [
          "posts"
        ],
        "summary": "Get Post Stats",
        "description": "Return aggregated statistics for the analytics dashboard.",
        "operationId": "get_post_stats_posts_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/posts/": {
      "post": {
        "tags": [
          "posts"
        ],
        "summary": "Create Post",
        "operationId": "create_post_posts__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "posts"
        ],
        "summary": "List Posts",
        "operationId": "list_posts_posts__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PostStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PostRead"
                  },
                  "title": "Response List Posts Posts  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/posts/{post_id}/submit/": {
      "post": {
        "tags": [
          "posts"
        ],
        "summary": "Submit Post For Review",
        "operationId": "submit_post_for_review_posts__post_id__submit__post",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Post Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/posts/{post_id}/publish/": {
      "patch": {
        "tags": [
          "posts"
        ],
        "summary": "Publish Post",
        "operationId": "publish_post_posts__post_id__publish__patch",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Post Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/posts/{post_id}": {
      "get": {
        "tags": [
          "posts"
        ],
        "summary": "Get Post",
        "operationId": "get_post_posts__post_id__get",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Post Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Health Check",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Check Health Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PostCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "content": {
            "type": "string",
            "minLength": 1,
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "title",
          "content"
        ],
        "title": "PostCreate"
      },
      "PostRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "status": {
            "$ref": "#/components/schemas/PostStatus"
          },
          "flagged_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Flagged Reasons"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "content",
          "status",
          "created_at"
        ],
        "title": "PostRead"
      },
      "PostStatus": {
        "type": "string",
        "enum": [
          "draft",
          "flagged",
          "approved",
          "published"
        ],
        "title": "PostStatus"
      },
      "ReasonCount": {
        "properties": {
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "reason",
          "count"
        ],
        "title": "ReasonCount"
      },
      "ReviewResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/PostStatus"
          },
          "flagged_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Flagged Reasons"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status"
        ],
        "title": "ReviewResponse"
      },
      "StatsResponse": {
        "properties": {
          "status_distribution": {
            "items": {
              "$ref": "#/components/schemas/StatusCount"
            },
            "type": "array",
            "title": "Status Distribution"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelineEntry"
            },
            "type": "array",
            "title": "Timeline"
          },
          "top_flagged_reasons": {
            "items": {
              "$ref": "#/components/schemas/ReasonCount"
            },
            "type": "array",
            "title": "Top Flagged Reasons"
          },
          "total_posts": {
            "type": "integer",
            "title": "Total Posts"
          },
          "approval_rate": {
            "type": "number",
            "title": "Approval Rate"
          },
          "flag_rate": {
            "type": "number",
            "title": "Flag Rate"
          }
        },
        "type": "object",
        "required": [
          "status_distribution",
          "timeline",
          "top_flagged_reasons",
          "total_posts",
          "approval_rate",
          "flag_rate"
        ],
        "title": "StatsResponse"
      },
      "StatusCount": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "status",
          "count"
        ],
        "title": "StatusCount"
      },
      "TimelineEntry": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "date",
          "count"
        ],
        "title": "TimelineEntry"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}